chore: implement grammar producers
Grammar rules use other grammar rules.

As a result, there can be an infinite loop when initializing the grammar rules.

Additionally, if only a few of the stored grammar rules are used, it can lead to a large inefficiency in memory usage.

Let's wrap the grammar rules in producer classes so that they are created only when necessary.